Propagation modeling API

Model database

UnderwaterAcoustics.addmodel! — Method

addmodel!(mtype)

Register model in the list of available models.

UnderwaterAcoustics.models — Function

models()
models(env::UnderwaterEnvironment)

List available models, optionally for a given environment.

Core interfaces

UnderwaterAcoustics.altimetry — Function

altimetry(env::UnderwaterEnvironment)::Altimetry

Get the altimetry for the underwater environment.

UnderwaterAcoustics.altitude — Function

altitude(alt:Altimetry, x, y)

Get water surface altitude at location (x, y). The nominal water surface is considered to have an altitude of zero. However, the water surface may not be flat, and the Altimetry provisions for variations of altitude around the nominal altitutde of zero.

UnderwaterAcoustics.arrivals — Function

arrivals(pm::PropagationModel, tx1::AcousticSource, rx1::AcousticReceiver)

Compute the arrivals from tx1 to rx1. Returns an array of Arrival structs.

UnderwaterAcoustics.bathymetry — Function

bathymetry(env::UnderwaterEnvironment)::Bathymetry

Get the bathymetry for the underwater environment.

UnderwaterAcoustics.check — Function

check(pm::Type{<:PropagationModel}, env::UnderwaterEnvironment)
check(pm::Type{<:PropagationModel}, env=missing)

Check if an propagation model is available, and can simulate the specified environment. Returns the environment if it can be simulated, or throws an error with a descriptive error message if it cannot be simulated.

This function is internally used by the propagation modeling toolbox to choose a model or offer a selection of models to the user.

UnderwaterAcoustics.depth — Function

depth(bathy:Bathymetry, x, y)

Get water depth at location (x, y).

UnderwaterAcoustics.eigenrays — Function

eigenrays(pm::PropagationModel, tx1::AcousticSource, rx1::AcousticReceiver)

Compute the eigenrays from tx1 to rx1. Returns an array of RayArrival structs.

UnderwaterAcoustics.environment — Function

environment(pm::PropagationModel)

Get the environment associated with the propagation model.

UnderwaterAcoustics.impulseresponse — Function

impulseresponse(arrivals, fs)
impulseresponse(arrivals, fs, ntaps; reltime, approx)

Convert a vector of arrivals to a sampled impulse response time series at a sampling rate of fs Hz. If ntaps is zero, the number of taps of the impulse response are chosen automatically.

If reltime is true, the impulse response start time is relative to the first arrival, otherwise it is relative to the absolute time. If approx is true, a fast algorithm is used to generate a sparse impulse response that assigns an arrival to the nearest sampling time.

UnderwaterAcoustics.location — Function

location(src::AcousticSource)
location(src::AcousticReceiver)

Get the location of an acoustic source or receiver as a 3-tuple (x, y, z).

UnderwaterAcoustics.maxdepth — Function

maxdepth(bathy::Bathymetry)

Get the maximum water depth.

UnderwaterAcoustics.noise — Function

noise(env::UnderwaterEnvironment)::NoiseModel

Get the noise model for the underwater environment.

UnderwaterAcoustics.nominalfrequency — Function

nominalfrequency(src::AcousticSource)

Get the nominal frequency of an acoustic source in Hz.

UnderwaterAcoustics.phasor — Function

nominalfrequency(src::AcousticSource)

Get the complex phasor representation (amplitude & phase) of a narrowband acoustic source at the nominal frequency.

UnderwaterAcoustics.rays — Function

rays(pm::PropagationModel, tx1::AcousticSource, θ::Real, rmax)

Compute the rays from tx1 launched at angle θ (or all angles in θ, if it is a vector). Returns an array of RayArrival datatypes. rmax is the maximum horizontal range in meters to track the rays over.

UnderwaterAcoustics.record — Function

record(src::AcousticSource, duration, fs; start=0.0)
record(noise::NoiseModel, duration, fs; start=0.0)
record(model::PropagationModel, tx, rx, duration, fs; start=0.0)

Make a recording of an acoustic source or ambient noise. The start time and duration are specified in seconds, and the recording is made at a sampling rate of fs Hz.

For an recording of an acoustic source, free space propagation is assumed, and the recording is made at a nominal range of 1 meter from the acoustic center of the source.

For a recording through a propagation model, tx and rx may be single AcousticSource and AcousticReceiver, or an array each.

UnderwaterAcoustics.recorder — Function

recorder(model::PropagationModel, tx, rx)

Create a recorder function that may be called later to make an acoustic recording of sources in tx at receviers rx. tx and rx may be single AcousticSource and AcousticReceiver, or an array each.

The recorder function may be called later with duration, fs, and optionally a start time. It functions in a similar way as the record() function.

Examples:

julia> rec = recorder(pm, tx, rx);
julia> s = rec(1.0, 44100.0; start=0.0);  # make a recording of 1 second at 44.1 kHz

UnderwaterAcoustics.reflectioncoef — Function

reflectioncoef(rm::ReflectionModel, f, θ)

Get complex reflection coefficient at frequency f Hz and incidence angle θ (w.r.t. the surface normal).

UnderwaterAcoustics.salinity — Function

salinity(env::UnderwaterEnvironment)

Get the salinity of the underwater environment.

UnderwaterAcoustics.seabed — Function

seabed(env::UnderwaterEnvironment)::ReflectionModel

Get the seabed reflection model for the underwater environment.

UnderwaterAcoustics.seasurface — Function

seasurface(env::UnderwaterEnvironment)::ReflectionModel

Get the sea surface reflection model for the underwater environment.

UnderwaterAcoustics.soundspeed — Function

soundspeed(ssp::SoundSpeedProfile, x, y, z)

Get sound speed at location (x, y, z). If a sound speed profile is range independent, x and y may be ignored. z is generally negative, since the sea surface is the datum and z-axis points upwards.

UnderwaterAcoustics.ssp — Function

ssp(env::UnderwaterEnvironment)::SoundSpeedProfile

Get the sound speed profile for the underwater environment.

UnderwaterAcoustics.transfercoef — Function

transfercoef(pm::PropagationModel, tx1::AcousticSource, rx1::AcousticReceiver; mode=:coherent)
transfercoef(pm::PropagationModel, tx1::AcousticSource, rx::AbstractArray{<:AcousticReceiver}; mode=:coherent)

Compute the complex transfer coefficients from tx1 to rx1 or all receivers in rx. The mode may be :coherent or :incoherent.

UnderwaterAcoustics.transmissionloss — Function

transmissionloss(pm::PropagationModel, tx1::AcousticSource, rx1::AcousticReceiver; mode=:coherent)
transmissionloss(pm::PropagationModel, tx1::AcousticSource, rx::AbstractArray{<:AcousticReceiver}; mode=:coherent)

Compute the transmission loss in dB from tx1 to rx1 or all receivers in rx. The mode may be :coherent or :incoherent.

Common models

UnderwaterAcoustics.AcousticReceiver — Method

AcousticReceiver(x, y, z)

Create an omnidirectional acoustic receiver at location (x, y, z).

UnderwaterAcoustics.AcousticReceiver — Method

AcousticReceiver(x, z)

Create an omnidirectional acoustic receiver at location (x, z).

UnderwaterAcoustics.AcousticReceiverGrid2D — Type

struct AcousticReceiverGrid2D{T} <: AbstractArray{UnderwaterAcoustics.BasicAcousticReceiver{T},2}

A 2D Cartesian grid of omnidirectional acoustic receivers.

UnderwaterAcoustics.AcousticReceiverGrid2D — Method

AcousticReceiverGrid2D(xmin, xstep, nx, zmin, zstep, nz)

Create a 2D Cartesian grid of omnidirectional acoustic receivers with nx × nz receviers starting (xmin, zmin) with step sizes xstep and zstep.

UnderwaterAcoustics.AcousticReceiverGrid3D — Type

struct AcousticReceiverGrid3D{T} <: AbstractArray{UnderwaterAcoustics.BasicAcousticReceiver{T},3}

A 3D Cartesian grid of omnidirectional acoustic receivers.

UnderwaterAcoustics.AcousticReceiverGrid3D — Method

AcousticReceiverGrid3D(xmin, xstep, nx, ymin, ystep, ny, zmin, zstep, nz)

Create a 3D Cartesian grid of omnidirectional acoustic receivers with nx × ny × nz receviers starting (xmin, ymin, zmin) with step sizes xstep, ystep, and zstep.

UnderwaterAcoustics.AcousticSource — Method

AcousticSource(x, y, z, f; sourcelevel, ϕ)

Create a narrowband acoustic source with frequency f Hz at location (x, y, z). The sourcelevel is in µPa @ 1m. A phase ϕ may be optionlly specified.

UnderwaterAcoustics.AcousticSource — Method

AcousticSource(x, z, f; sourcelevel, ϕ)

Create a narrowband acoustic source with frequency f Hz at location (x, y, z). The sourcelevel is in µPa @ 1m. A phase ϕ may be optionlly specified.

UnderwaterAcoustics.BasicAcousticReceiver — Type

struct BasicAcousticReceiver{T} <: AcousticReceiver

Omnidirectional acoustic receiver.

UnderwaterAcoustics.BasicAcousticReceiver — Method

BasicAcousticReceiver(x, y, z)

Create an omnidirectional acoustic receiver at location (x, y, z).

UnderwaterAcoustics.BasicAcousticReceiver — Method

BasicAcousticReceiver(x, z)

Create an omnidirectional acoustic receiver at location (x, z).

UnderwaterAcoustics.ConstantDepth — Type

struct ConstantDepth{T} <: Bathymetry

Bathymetry with constant depth.

ConstantDepth(depth)

Create a constant depth bathymetry.

UnderwaterAcoustics.FlatSurface — Type

struct FlatSurface <: Altimetry

Altimetry for a flat surface with constant altitude of zero.

UnderwaterAcoustics.IsoSSP — Type

struct IsoSSP{T} <: SoundSpeedProfile

Isovelocity sound speed profile.

IsoSSP(c)

Create an isovelocity sound speed profile with sound speed c.

UnderwaterAcoustics.MunkSSP — Type

struct MunkSSP <: SoundSpeedProfile

Munk sound speed profile.

UnderwaterAcoustics.NarrowbandAcousticSource — Type

struct NarrowbandAcousticSource{T1, T2, T3, T4} <: AcousticSource

Narrowband acoustic source.

UnderwaterAcoustics.NarrowbandAcousticSource — Method

NarrowbandAcousticSource(x, y, z, f; sourcelevel, ϕ)

Create a narrowband acoustic source with frequency f Hz at location (x, y, z). The sourcelevel is in µPa @ 1m. A phase ϕ may be optionlly specified.

UnderwaterAcoustics.NarrowbandAcousticSource — Method

NarrowbandAcousticSource(x, z, f; sourcelevel, ϕ)

Create a narrowband acoustic source with frequency f Hz at location (x, z). Thesourcelevelis in µPa @ 1m. A phaseϕ` may be optionlly specified.

UnderwaterAcoustics.Pinger — Method

Pinger(x, y, z, f; sourcelevel, phase, duration, start, interval, window)

Create a pulsed narrowband acoustic source with frequency f Hz at location (x, y, z). Additional parameters that may be specified:

sourcelevel in µPa @ 1m (default 180 dB)

phase of the narrowband signal (default 0)

duration of the pulse in seconds (default 20 ms)

start time of one of the pulses in seconds (default 0)

pulse repetition interval in seconds (default 1 second)

window type (from DSP.jl) (default nothing)

UnderwaterAcoustics.Pinger — Method

Pinger(x, z, f; sourcelevel, phase, duration, start, interval, window)

Create a pulsed narrowband acoustic source with frequency f Hz at location (x, z). Additional parameters that may be specified:

sourcelevel in µPa @ 1m (default 180 dB)

phase of the narrowband signal (default 0)

duration of the pulse in seconds (default 20 ms)

start time of one of the pulses in seconds (default 0)

pulse repetition interval in seconds (default 1 second)

window type (from DSP.jl) (default nothing)

UnderwaterAcoustics.RayleighReflectionCoef — Type

struct RayleighReflectionCoef{T1, T2, T3} <: ReflectionModel

Reflection model for a surface with a Rayleigh reflection coefficient.

RayleighReflectionCoef(ρᵣ, cᵣ, δ)
RayleighReflectionCoef(ρᵣ, cᵣ)

Create a reflection model for a surface with a Rayleigh reflection coefficient with relative density ρᵣ, relative sound speed cᵣ, and dimensionless attentuation δ. If attentuation δ is unspecified, it is assumed to be zero.

See reflectioncoef() for more details.

UnderwaterAcoustics.RedGaussianNoise — Type

struct RedGaussianNoise{T} <: NoiseModel

Ambient noise model with Gaussian noise with a 1/f² power spectral density.

RedGaussianNoise(σ)

Create an ambient noise model with variance σ² and 1/f² power spectral density.

UnderwaterAcoustics.ReflectionCoef — Type

struct ReflectionCoef{T<:Number} <: ReflectionModel

Reflection model for a surface with a constant reflection coefficient.

ReflectionCoef(coef)

Create a reflection model for a surface with a constant reflection coefficient coef.

UnderwaterAcoustics.SampledAcousticSource — Type

struct SampledAcousticSource{T1, T2, T3<:AbstractArray} <: AcousticSource

Acoustic source transmitting a sampled signal.

UnderwaterAcoustics.SampledAcousticSource — Method

SampledAcousticSource(x, z, sig; fs, frequency)

Create a sampled acoustic source transmitting signal sig at location (x, z). The samples are assumed to be in µPa @ 1m from the source.

Additional parameters that may be specified:

fs is the sampling rate of signal sig

nominal frequency in Hz (nothing to auto-estimate)

UnderwaterAcoustics.SampledAcousticSource — Method

SampledAcousticSource(x, y, z, sig; fs, frequency)

Create a sampled acoustic source transmitting signal sig at location (x, y, z). The samples are assumed to be in µPa @ 1m from the source.

Additional parameters that may be specified:

fs is the sampling rate of the signal sig

nominal frequency in Hz (nothing to auto-estimate)

UnderwaterAcoustics.SampledAltitude — Type

struct SampledAltitude{T1, T2, T3} <: Altimetry

Altimetry based on altitude samples.

UnderwaterAcoustics.SampledAltitude — Method

SampledAltitude(x, altitude)
SampledAltitude(x, altitude, interp)

Create an altimetry given discrete altitude measurements at locations given in x. interp may be either :linear or :smooth, and defaults to :linear if unspecified.

UnderwaterAcoustics.SampledDepth — Type

struct SampledDepth{T1, T2, T3} <: Bathymetry

Bathymetry based on depth samples.

UnderwaterAcoustics.SampledDepth — Method

SampledDepth(x, depth)
SampledDepth(x, depth, interp)

Create a bathymetry given discrete depth measurements at locations given in x. interp may be either :linear or :smooth, and defaults to :linear if unspecified.

UnderwaterAcoustics.SampledSSP — Type

abstract type SampledSSP <: SoundSpeedProfile

Sound speed profile based on discrete measurements.

UnderwaterAcoustics.SampledSSP — Method

SampledSSP(depth, c)
SampledSSP(depth, c, interp)
SampledSSP(range, depth, c)
SampledSSP(range, depth, c, interp)

Create a sound speed profile based on measurements at discrete depths. interp may be either :linear or :smooth, and defaults to :linear if unspecified. :smooth is currently only supported for range-independent sound speed profiles.

UnderwaterAcoustics.SampledSSP1D — Type

struct SampledSSP1D{T1, T2, T3} <: SampledSSP

Sound speed profile based on measurements at discrete depths.

UnderwaterAcoustics.SampledSSP2D — Type

struct SampledSSP2D{T1, T2, T3, T4} <: SampledSSP

Sound speed profile based on measurements at discrete depths and ranges.

UnderwaterAcoustics.SurfaceLoss — Type

struct SurfaceLoss{T} <: ReflectionModel

Reflection model for a water surface affected by wind.

SurfaceLoss(windspeed)

Create a reflection model for a surface affected by wind. windspeed is given in m/s.

UnderwaterAcoustics.UnderwaterEnvironment — Method

UnderwaterEnvironment(; altimetry, bathymetry, ssp, salinity, seasurface, seabed, noise)

Create an underwater environment.

Propagation models

UnderwaterAcoustics.PekerisRayModel — Type

struct PekerisRayModel{T} <: PropagationModel{T}

A fast differentiable ray model that only supports isovelocity constant depth environments.

PekerisRayModel(env, rays)

Create a Pekeris ray propagation model with a maximum of rays ray arrivals.

UnderwaterAcoustics.RaySolver — Method

RaySolver(env; nbeams, minangle, maxangle, ds, atol, rugosity, athreshold, solver, solvertol)

Create a RaySolver propagation model.

UnderwaterAcoustics.Bellhop — Type

struct Bellhop{T} <: PropagationModel{T}

A propagation model based on an external FORTRAN Bellhop executable.

UnderwaterAcoustics.Bellhop — Method

Bellhop(env; gaussian=false, debug=false)
Bellhop(env, nbeams, minangle, maxangle, gaussian, debug)

Create a Bellhop propagation model.